Make cropping work consistently
