+++ .eslintrc
@@ -187,7 +187,7 @@
 		"operator-assignment": [1, "never"], // require assignment operator shorthand where possible or prohibit it entirely (off by default)
 		"padded-blocks": [1, "never"], // enforce padding within blocks (off by default)
 		"quote-props": [1, "as-needed"], // require quotes around object literal property names (off by default)
-		"quotes": [1, "single"], // specify whether double or single quotes should be used
+		"quotes": [0, "single"], // specify whether double or single quotes should be used
 		"semi": [1, "always"], // require or disallow use of semicolons instead of ASI
 		"semi-spacing": [1, {
 			"before": false,
