+++ create-release.sh
@@ -26,7 +26,7 @@ git tag -a $VERSION -m "Public release of version ${VERSION}"
 git push > /dev/null
 git push --tags > /dev/null
 
-echo "Version ${VERSION} was succesfully released. On Github you can complete the release: https://github.com/inventid/live-image-resize/releases/tag/${VERSION}"
+echo "Version ${VERSION} was successfully released. On Github you can complete the release: https://github.com/inventid/live-image-resize/releases/tag/${VERSION}"
 git checkout develop > /dev/null
 exit 0
